@todo, create, does>, API rename, \ comments, ...
Some @todo notes have been added to "libforth.h"

A working "create" and "does>" have been added, although they need a lot of
work doing to them.

Function in the API have been renamed.

Traditional forth '\' comments have been added, '#' as a comment has been
removed (although it acts as one if it is the first character of a file being
read in).