removed doxygen stuff, new hexdump util, more words
The doxygen stuff has been removed, it was unneeded.

A new hexdump utility has been added for helping analyze the core dumps.

More words continue to be added.

STDIN can be read from when executing from a file, allowing hexdump to be made,
this allows for some simple utilities to be made.