hexdump, more pointer stuff, better i/o, ...
More pointers that were in "struct forth" have been moved so as to be accessible
by the Forth interpreter which has allowed a simplification of the I/O
mechanism, and would EVALUATE to be coded. This is actually quite a big change
to the interpreter as it now means the size of "forth_cell_t" (which specifies
the Forth virtual machines word size) needs to be large enough to store a
pointer in it (or equal to uintptr_t in size).