Improved argument processing, man pages. Performance Hit?
The processing of command line arguments has been greatly improved, allowing
for more functionality. An "-e" option, for evaluating a string, might be
planned in future releases. The man pages have been updated to reflect these
changes.

Somewhere since in the last few commits the performance of the interpreter has
taken a hit, why this is needs to be ascertained and corrected.