Better decompilation and debugging, memory options, more...
Decompilation and debugging has been improved, "see" nearly works correctly
but there is still more to do. Decompilation is quite difficult, as there are
quite a few corner cases that need to be sorted out, it might be impossible to
do in general.

More words for debugging has been added, such as "step", and the way
hexadecimal values are printed out has been changed so things align up better.

I am worried that my forth code is not optimal, it is not readable as it could
be and I should refactor a lot of it so it is more "forth like". This could be
done by splitting long definitions of words into small chucks, and getting more
used to the language, finding out how other forths do things would help as
well.

There is now a "memory" option for specifying how much memory the VM should be
able to use.

Some words have been moved around, commented and improved upon already, but
there is a lot to do still, this is before even adding in proper string
literals, operating system interfaces, and a whole host of other words.