case, [literal], comments
Case statements have been added to Forth, with a few different variants
available in the comments, they need to be integrated and used within the
interpreter in various places.

"[literal]" has been added as Forth word and used, but it probably needs a
better name.

Comments have been added to the run length encoder, but more should be added,
as well as man pages, comments have been added elsewhere in the Forth code as
well for things that need to be done with the interpreter.