Unit tests, bug fixes, more forth code, ...
The unit tests have been improved with
	- more tests
	- a silent mode
	- they are now part of more of the
	  build

The makefile has been improved, more things execute the unit tests

rle.c has had a warning about clobbering resolved (although it would not have
been a problem in this case).

More forth code has been added, the start of a debug command interpreter has
been added, but changes to the interpreter are needed in order to complete it,
and the word itself improved.