Added BNF (todo), fixed bug in s", added jump table tests
Added a one screen BNF parser, which will not work at the moment, but that is
on the To-Do list, it is quite impressive - it implements a BNF parser in such
a small space.

Fixed a bug in s", where it wrote to word address instead of a character
address.

A few extra words have been added.

Some jump table examples have been added.