File access methods, bug detected, misc.
File access methods have been added, this is a big change to the interpreter.
This will allow more functionality to be put into "forth.fth", allowing more
complex Forth programs to be made. There is still a lot of work to be done
around them.

A bug has been detected in the 32-bit version, the unit tests in "unit.fth" do
not run correctly when the "unit(" word is used.