Line editor, documentation generation, logging
The documentation generation system has been improved, but there is still more
work to be done, the manual page "forth.1" has been removed, it is instead
generated from the "readme.md" file. In the future "libforth.3" should be
generated from the header "libforth.h", which will require modification to the
header file and the "convert" program. Some of the changes to the "convert"
program have already been made in this commit.

The logging with libforth has been improved with some wrappers around printf,
making the error messages displayed more consistent.

An optional line editor has been included, this needs to be compiled into the
program, but by default it is not included. This uses the libline library.