Improved documentation, X-Macro, removed instructions, ...
The documentation and the documentation generation system has been improved,
the manual pages can be generated from markdown files, or from C code that is
specially formatted.

Various instructions that are no longer needed have been removed; such as
signed comparison operators and EMIT.

An X-Macro linking instructions with their names has been made, along with a
help text for each instruction which is not used at the moment.