Improved unit tests, bug fixes, new words
More unit tests have been added, but there is still an incredible amount of
work to be done with them.

A few new words have been added, such as abs and signum. As well as the word
"0", "1", and a few other numbers (to save on space).

A bug has been fixed in the signed comparison operations.