Preparing new unit test framework, better looping, ...
A few new words have been added like "equal" and "ndrop", the main changes
however are the new unit test frame work and the rewriting of do...loop.

A better, more advanced, unit test framework has nearly been finished, it could
be integrated into the "forth.fth" instead of a separate file so tests are
always run (with a verbosity option for whether error messages are printed out
or not).

do...loop has also been rewritten, although there is more to do if "?do" is to
be accommodated. It is clearer and better factored.