Added new words, catch/throw and others.
catch and throw have been added but they are not used at the moment, they
should be used as much as is possible and further integrated into libforth. A
few virtual machine instructions have been added to accommodate this.

A few other words have also been added to the interpreter.