Added RLE decompressor in Forth, Error generator
A Run Length Encoder decompressor has been added, the compressor has not been
made yet, it has been tested and it works from the output of another RLE
compressor written in C.

A file for generating the strings corresponding to error numbers has been added
to the repository. This is non-portable, so it probably will be removed,
strerror could be used instead...

There have been other minor changes as well:
* A colorize variable controls the output of the ANSI color codes, color can
output now be turned on or off.
* "?if" has been added, a non-destructive if, words like "-if" should also be
added.
* A list of error codes has been added, these should be use throughout
"forth.fth", which will take time to do, it will also require better
interaction from the error handler in "libforth.c".
* Minor documentation changes and updates.