Reorganized code, hexdump, rewriting blocks, ...
A hexdump utility has been added which can dump a file to screen (or two a file
if the output device is redirected).

"mini.c" contains a small example program with a much simpler main() than the
one in libforth, but it still loads and saves a core file.

The code has been reorganized and moved around, the block primitives have been
removed in favor of the built in file utilities.

The blocks code has been removed, and is in the process of being rewritten.