Forth as a Linux Kernel Module - Initial Commit
This branch is for the development of Forth as a Linux Kernel Module, the
initial commit is far from working. It creates a device, it does not crash and
does some initialization, but there is still a lot of work to do to get this
working.

The idea is that this Forth could be used to debug the Kernel, and just for
fun.