Kernel module works
Whilst there is still a lot of work to do, the Forth interpreter now works, a
test program can exercise the module. This is a good starting point from which
the functionality of the module can be expanded.

A test program, "test.c", can be used to test the module.