API improvements
The libforth API has been improved, more functionality has been exposed in the
header from internal utility functions, and two new functions relating to the
invalidation status of a Forth environment have been added.

Forth loading/saving to memory now works on bytes and includes the header. The
utilities under util/ have been updated to reflect this.

A stub file has been added for a compiler which will target the virtual machine
provided in libforth.c

The file handles on Windows platforms for stdin, stderr and stdout should be
fixed (turned into reading/writing in binary mode), this needs testing on a
Windows machine.

TODO:
* The libforth.h needs updating to reflect these changes.
* Also the documentation generated by "make libforth.pdf" needs the header adding
as an appendix, as it is incomplete without it.
* More unit tests need writing
* Most of the functionality for main_forth needs moving into main, the
core versions of main_forth needs to be made simpler.