Finished feature, integrated unit tests, main()
The feature branch has now been completed, a binary called "libforth" can be
produced which includes a previously produced forth core. The build process is
now:

* built the "forth" executable
* run "forth" on "forth.fth" to produced a core file
* turn this into a C file
* recompile the "forth" executable with the new C file to produce "libforth"

The "libforth" now includes the contents of "forth.fth".

There may be problems with alignment with the generated file "core.gen.c" which
will have to be investigated. This might mean execution failure on some
architectures.

"main_forth" has been replaced by a simpler alternative, and the functionality
moved to "main.c". The API has had to be changed to accommodate this, more
functionality is now exported.