Allocate, free, memory operations.
Allocate and free have been added as have memory operations that can access
memory outside of the Forth core. At the moment the Forth words '@' and '!'
cannot access memory outside of the Forth core, and are biased so addresses
start at first memory address of the Forth core. Memory operations such as
memory-compare and memory-copy can be used to interact with memory outside of
the Forth core.

This will eventually be used for processing command line arguments passed into
the Forth interpreter.