Reorganized project, changed command line options, ...
Removed compression library submodule and unneeded code in "util/".

The command line options now make more sense, or have been rationalized
slightly.

Converted "core2c.c" to a set of forth words, changed build system so when
building "libforth" it uses the forth word to do this. This process has been
explained in the "readme.md" document. It is still experimental.

Added 'wc' and 'nos1+' words.

Added printing of stack trace on unix systems with backtrace if SIGABRT is
called.