Miscellaneous changes, strange bug "fix", ...
There is a strange bug with a strange fix in "forth.fth", consult line 750 for
more detail. I have applied a temporary solution at the moment, but I will try
to replicate the bug later and fix it.

There have been a few other small changes, to help text and to build options,
the latter may be reverted or changed - I am not too fussed about the compiler
options used.