Added documentation, fixed static target, convert
The "convert" requires gawk, mawk tends to be installed by default on a fair
few distributions, this has not been made explicit in the script. This
gawk-isms should really be removed, however it is not really a high priority,
so the script is likely to stay as is.

How to add "magic" so that Forth core files can be recognized as been added as
a description to "readme.md". These "magic" files are used by the file utility
to determine file type and various attributes of the files they analyze.

"-static" has been added to the flags of the static target, which was missing.