Semi-Working commit - Compiling Bit
This is the first commit of a new feature of the virtual machine, READ will use
a "Compiling-bit" instead of COMPILE/RUN virtual machine instructions to
determine whether a word should be executed or compiled into the dictionary
when the interpreter is in the compile mode.

This will require reworking some tricky Forth code that has to know about the
interpreter internals, and will require rewriting and rewording a lot of the
documentation.

This commit is semi-working, as the commit messages says, the mechanism works
but a lot of the code in "forth.fth" does not. At the moment it is not clear
why this is (although this was expected and is part of the process of adapting
the virtual machine for this).

This change was done for multiple reasons:

1) To allow the Forth interpreter to be more compliant.
2) To save space (this was the main reason).
3) To simplify the internals of the interpreter.

It will allow the "immediate" word to be used in the definition of an immediate
word like this:

	: xxx ... ; immediate

Instead of like this:

	: xxx immediate ... ;

As it has currently been done.

More @todos have been added as well, and *some* of the documentation has been
updated (or just deleted).

The code in "forth.fth" words up until where it has been commented that it does
not, upon which it exits instead of trying to execute non-functional code.