Added 'r.s', moved words around, Windows unit test
The 'r.s' has been added, which prints out the return stack.

'words' optionally colorizes it's output, which allows a quick and easy way to
determine if a word is built-in or an immediate word.

Tests were failing under Windows because of a brain dead implementation of
'tmpfile()' on that platform.

Some words have been moved around as well.