Signal handling, words, explanation, ...
As usual there have been a few minor changes throughout, the major ones are
the inclusion of a primitive form of signal handling - which should be improved
upon later. The explanation of the internals in "libforth.c" has also been
updated.

The words 'sh', 'r13' and 'r13-type' have also been added.