Rename MISC->CODE, fence, constants X-Macro, bias, ...
Documentation has been improved, reflecting the different mechanism 'see' uses
to differentiate between an exit part of a word and an exit terminating the
word inserted by ';'.

The MISC field has been renamed to the more logical CODE field. This is a
better description of the former MISC field.

The word 'fence', and mechanism for it, has been added to 'forth.fth'.

An X-Macro for adding constants has been added to the interpreter, this makes
adding constants a data driven process.

Signals are biased now, errnos will have to be biased as well.