Fixed most doxygen warnings, errno biasing.
Errnos that reach the interpreter should be biased so they can be used with
THROW in a standards (ANS Forth) compliant manner. This does add some
complication and creates an impedance mismatch between the C environment and
the Forth environment, but it requires little effort.

Most doxygen warnings have been fixed, the others (such as @license and @email
being unknown commands) would have to be fixed with a custom doxygen
configuration file.