Improved decompiler, ...
The decompiler has been improved but there are still many changes that need to
be made until it is useful. The 'decompile' word has been factored into two
words, one of which is not hidden and can be used to manually decompile
addresses (for interactive use).

A few new words have been added as well ".h", "+-", "sign-bit", as well as some
comments.