Constants, see, ...
There is now a better definition of constants backed by a new instruction that
pushes the next location after the CONST instruction, not the next location in
the instruction stream, which both saves spaces and is faster.

The decompiler has been improved, but still needs *a lot* of work. CONST also
needs handling.