Blocks, comments, Caesar cipher, ...
This commit only changes Forth code, it adds the block layer which had been
removed in a few commits prior to this one. It needs work, but seems okay.

Various comments have been expanded on.

A generalization of the Caesar cipher has been added to "forth.fth", which is
obviously just a toy. It can be used for rot-13 encoding text.