Optimized find (not used), more constants, ...
A faster version of find has been added as an option (selectable by compiling
libforth with USE_FAST_FIND defined), it changes the dictionary to use a self
organizing list (by using transposition). It is an option as it currently does
not interact with the forth word "marker".

More words have been turned into constants where they can be, and the constants
have been moved to the beginning of "forth.fth".

The documentation of DOER/MAKE has been improved.

"code>pwd" has also been greatly simplified which speeds up the decompiler a
lot, I believe it to be correct and cannot think of a case where it would fail,
however the decompiler is quite brittle anyway.