Optional signal handling, debug, ...
There are been a fair few miscellaneous changes made throughout the
interpreter. One is making the installation of the signal handler optional,
this was annoying me when a program got stuck in an infinite loop and did not
respond (which will not happen if the signal handler made the Forth interpreter
throw an exception). This change affects "main.c".

Some codes indentation has changed and a register has been renamed in
"libforth.c".

An experimental block editor has been laid out in "forth.fth", it is currently
commented out.

The "debug" word has been improved and is now useful. This word can be inserted
into the source code of a Forth word to debug the interpreter state.

Moving forward, operating system specific code may have to be added into a
separate C file and integrated into the interpreter. I am not sure how I feel
about this, however this may be the only way for the interpreter to reach the
next level.

There is quite a lot to do to make this Forth more standards compliant, I am
slowly working on it.

Some commented out code for the words FORTH and VOCABULARY have been added, but
they are not very useful without using Forth word-lists.

Other changes included commenting things, moving Forth words around, and other
tiny changes.