Added a primitive block editor, ...
A primitive block editor has been added to "forth.fth", it allows the editing,
saving and evaluation of Forth blocks. There is still plenty of work to do
to make it more usable, making it revealed some things that are not standards
compliant in Forth that will need changing. Mostly in how the primitive
"evaluator" does things.

There have been a few other changes as well.

* '\n' has been renamed to 'nl.
* A lot more @todo comments have been added.
* Some other things have been renamed and hidden in different areas.