Coding standards, foreach, ?comp, ?exec, case, ...
Coding standards for both Forth and C have been added to "readme.md", there is
still more to document, but it outlines most points.

The 'foreach' word set has been greatly improved on, making for a powerful
looping mechanism.

The words '?comp' and '?exec' have been added, and are used for some control
structures, but could be used for more of them if they were to be redefined
later.

Unit tests have been moved to around and added to "unit.fth", also if an
exception that is uncaught occurs during the tests the tests now fail, instead
of continuing on.