Formatting, line editor completion callback, words
Two new functions have been added to the libforth library, one for returning a
list of all defined words within the interpreter, another for freeing the
returned list. This is used for the next, optional, feature - adding
completions for partially typed in words when using the line editor. The
feature is still fully not ready. The default behavior is to add all the words
as completions, and there appears to be a memory problem that valgrind detects
when using the libline library as a line editor - it is not clear where the
problem lies just yet. A string duplication has also been added, forth_strdup.

The formatting of "forth.fth" has been improved, there is still a lot of work
to do.