Improved error output, fixed NUL in evaluate.
There have been some minor changes, such as using ['] instead of ' in word
definitions, and some others. The major changes are improved error output, the
line number is displayed when an error has occurred, and evaluate no longer
requires a NUL terminated string in order to evaluate a block of text.